test(hooks): always rebuild dist under repo-scoped lock for symlink tests
Extract shared cli-dist-build-coordination helper so symlink entrypoint
regressions always run incremental tsc -b (no stale-dist skip) and parallel
dist readers wait for the build lock before executing dist artifacts.
